+++ src/vumi2/applications/turn_channels_api/__init__.py
@@ -1,0 +1,3 @@
+from .turn_channels_api import TurnChannelsApi
+
+__all__ = ["TurnChannelsApi"]